requirements.txt
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
dash==1.10.0
dash-core-components==1.9.0
dash-html-components==1.0.3
dash-renderer==1.3.0
dash-table==4.6.2
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numexpr==2.7.1
numpy==1.18.2
orca==1.5.3
pandas==1.0.3
plotly==4.6.0
psutil==5.7.0
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
retrying==1.3.3
six==1.14.0
tables==3.6.1
toolz==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==1.0.1
wrapt==1.11.2
